FEAT: tests improved

+++ services/backend/build/init.sh
@@ -4,6 +4,12 @@ set -eu
 
 mkdir -p /backend/DB/website
 
+if [ "$APP_MODE" = "test" ]; then
+    echo "Running Tests..."
+    exec conda run --no-capture-output -n backend python /backend/manage.py test tests
+    #tests.test_userprofile
+    exit 0
+fi
 
 rm -f /backend/DB/website/db.sqlite3
 find /backend -path "*/migrations/0*" -delete
@@ -12,11 +18,6 @@ conda run -n backend python /backend/manage.py makemigrations
 conda run -n backend python /backend/manage.py migrate
 conda run -n backend python /backend/manage.py collectstatic --noinput
 
-if [ "$APP_MODE" = "test" ]; then
-    echo "Running Tests..."
-    exec conda run --no-capture-output -n backend python /backend/manage.py test
-    exit 0
-fi
 
 conda run -n backend bash -c "python /backend/manage.py shell < /backend/seed.py"
